Update fixture and tests for non-iocage jail
Without this change, the provider attempts to parse the data from
'iocage list' even when non-iocage-managed jails are present.  Due to a
difference between the way iocage jails and non-iocage jails are
represented, the resutls are incorrect, and cause problems with jail
detection.

This work updates the fixtures to contain such an example (during a
poudriere build) and builds a regex to match only the iocage jails.
Currently we throw away the matches and only test that we actually have
the correct data.