Fix regex match to allow for zeros in JID
Without this change, Jail IDs that contain a zero are not matched
causing a duplicate jail to be created.  This quickly gets out of hand,
as the following run it will still not be matched to the duplicate tag
name, and iocage will fail causing puppet to create yet another jail.