Add user_data support
This work adds a new parameter to support passing in user_data content
upon first boot.  This is implemented in the form of creating a
temporary file passed to the execute() method in the Puppet utilities.
The file relies on the default mode of 0600 for the creation, and is
removed once executed.