Improve provider tests
Currently, in order to make simple changes to the module, it is not clear that
the correct change is being made, and there are not tests in place to illustrate
a change in behavior.  Here we begin to improve our provider tests so that we
may have more confidence in future changes.