Fix bug in down state where create is called
Without this change, we call create while trying to bring down a jail that is
present.  This raised the issue that there were no tests for the flush method,
so here we include some tests and also refactor the flush method to avoid use of
the @property_flush hash.  It was a style that I was trying out for a while, and
doesn't actually make sense in this case.  The structure of the flush method has
been cleaned up to make a bit more sense.