pkglist, step 1: add param, ensure nothing breaks
we add the pkglist parameter and it's appropriate pkglist= provider
function. This function will use a helper (`update`), which in turn will
reuse the destroy & create behaviour. Since that is currently stashed
away in the body of `flush`, we have to extract it.

This commit verifies that we didn't break any of the previous behaviour.