Updates as part of testing
Here are a bunch of changes that were needed to get this version of the code
live in my environment.  Much of this was needed as the upstream iocage was very
unstable for a period of time, but seems to be stabilizing somewhat, anecdotally
observed by github activity.

There are most certainly still issues here, but the core functionality does seem
to work for me reasonably well, and I personally rely on this module to get my
work done.  What follows are brief descriptions of the changes.

Update package name.  The name of the iocage package has changed recently to
reflect the rewrite in python.  Furthermore, it is unclear if the prefix of py36
will remain, as is hardcoded here.  As such, we allow tuning the package name.

Improve validation on release property

Create and destroy only when necessary

Correct parameter order on resource

Fixes for rubocop

Include more debug

Adjust address logic

Check for the property_hash ensure before starting

Check discovered state for creation

Start a jail only if necessary

Wrap the address in quotes

Add the properties individually

Skip built in jail properties

Set boot state

Use pipe for executing from stdin

Better iocage path

Bring back legacy spec fixuture.  This was deleted as part of prior work due to
the expectation that we would retire the legacy provider as part of this work.
With the length of time that this work has been outstanding, there have been
commits here against the legacy provider that I would like not to loose so as to
avoid breaking folks who have started using this module recently.  Versions are
cheap, and so we might seek to pull out the legacy provider as its own effort,
followed by cutting a new release.