Set the property only if required
Without this change, all properties are set each application.  In my
environment, this appears to be due to a recent change where devfs_rulesets
don't appear to match the defaults, and are detected as a change by puppet,
which then are changed back to the correct values, though some of the values may
already be set to the desired value.  Here we ensure that the values that match
the current value are not attempted to be set.  This avoids unnecessary calls to
the iocage utility, and also skips attempts to modify settings that are not
settable while a jail is running.